Target specific code coverages
This adds makefile targets that calculate the code coverages per target,
including the main file of the target. An attempt is made to exclude the
packages that are not part of the target.

The old coverage target is changed to be a plain html generation target.
The new excluded-coverage target can be used with the default
COVERAGE_EXCLUDE variable to get a blind average coverage of the whole
codebase, fwiw.