Fix test container UID and GID usage
The test container couldn't be run if the UID and GID of the user were
something else than 1000:1000. This change allows for using
current user's UID and GID while building the test container and
running the tests.

Also the safe.directory git config is removed as it becomes redundant
when the UID and GID of the home directory are set to match the running
user. The need for UID and GID matching arises from the use of host
directories in the container.